requirements.txt
gunicorn
Django
django-cors-headers
djangorestframework
python-dotenv
